propchain_monitoring::monitoring

Struct MonitoringContractRef

pub struct MonitoringContractRef { /* private fields */ }

Implementations

impl MonitoringContractRef

pub fn new() -> CreateBuilder<<MonitoringContract as ContractEnv>::Env, Self, Unset<<<MonitoringContract as ContractEnv>::Env as Environment>::Hash>, Set<LimitParamsV2<<MonitoringContract as ContractEnv>::Env>>, Unset<<<MonitoringContract as ContractEnv>::Env as Environment>::Balance>, Set<ExecutionInput<EmptyArgumentList>>, Unset<Salt>, Set<ReturnType<Self>>>

Deploys the monitoring contract. The caller becomes admin.

pub fn set_health_status( &mut self, status: HealthStatus, ) -> Result<(), MonitoringError>

Manually override the stored health status. Admin only.

pub fn try_set_health_status( &mut self, status: HealthStatus, ) -> MessageResult<Result<(), MonitoringError>>

Manually override the stored health status. Admin only.

pub fn set_alert_config( &mut self, alert_type: AlertType, threshold_bips: u32, active: bool, ) -> Result<(), MonitoringError>

Configure an alert type. threshold_bips = 0 means “use default”. Admin only.

For HighErrorRate: threshold_bips is the error-rate trigger level. For SystemDegraded: threshold_bips is ignored.

pub fn try_set_alert_config( &mut self, alert_type: AlertType, threshold_bips: u32, active: bool, ) -> MessageResult<Result<(), MonitoringError>>

Configure an alert type. threshold_bips = 0 means “use default”. Admin only.

For HighErrorRate: threshold_bips is the error-rate trigger level. For SystemDegraded: threshold_bips is ignored.

pub fn get_alert_config(&self, alert_type: AlertType) -> AlertConfig

Returns the current configuration for a given alert type.

pub fn try_get_alert_config( &self, alert_type: AlertType, ) -> MessageResult<AlertConfig>

Returns the current configuration for a given alert type.

pub fn subscribe_alerts( &mut self, subscriber: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Add an account to the alert subscriber list. Admin only.

pub fn try_subscribe_alerts( &mut self, subscriber: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Add an account to the alert subscriber list. Admin only.

pub fn unsubscribe_alerts( &mut self, subscriber: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Remove an account from the alert subscriber list. Admin only.

pub fn try_unsubscribe_alerts( &mut self, subscriber: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Remove an account from the alert subscriber list. Admin only.

pub fn get_alert_subscribers( &self, ) -> Vec<<<MonitoringContract as ContractEnv>::Env as Environment>::AccountId>

Returns the list of registered alert subscribers.

pub fn try_get_alert_subscribers( &self, ) -> MessageResult<Vec<<<MonitoringContract as ContractEnv>::Env as Environment>::AccountId>>

Returns the list of registered alert subscribers.

pub fn add_reporter( &mut self, reporter: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Authorize an external account or contract to call record_operation. Admin only.

pub fn try_add_reporter( &mut self, reporter: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Authorize an external account or contract to call record_operation. Admin only.

pub fn remove_reporter( &mut self, reporter: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Revoke a previously authorized reporter. Admin only.

pub fn try_remove_reporter( &mut self, reporter: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Revoke a previously authorized reporter. Admin only.

pub fn is_authorized_reporter( &self, account: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> bool

Returns whether account is an authorized reporter.

pub fn try_is_authorized_reporter( &self, account: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<bool>

Returns whether account is an authorized reporter.

pub fn pause(&mut self) -> Result<(), MonitoringError>

Pause the contract, blocking new operation recordings and snapshots. Admin only.

pub fn try_pause(&mut self) -> MessageResult<Result<(), MonitoringError>>

Pause the contract, blocking new operation recordings and snapshots. Admin only.

pub fn resume(&mut self) -> Result<(), MonitoringError>

Resume a paused contract and restore the health status to Healthy. Admin only.

pub fn try_resume(&mut self) -> MessageResult<Result<(), MonitoringError>>

Resume a paused contract and restore the health status to Healthy. Admin only.

pub fn get_admin( &self, ) -> <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId

Returns the admin account.

pub fn try_get_admin( &self, ) -> MessageResult<<<MonitoringContract as ContractEnv>::Env as Environment>::AccountId>

Returns the admin account.

pub fn transfer_admin( &mut self, new_admin: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Transfer admin rights to a new account. Admin only.

pub fn try_transfer_admin( &mut self, new_admin: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Transfer admin rights to a new account. Admin only.

pub fn register_health_contract( &mut self, contract: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Register a contract for health-check aggregation. Admin only.

pub fn try_register_health_contract( &mut self, contract: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Register a contract for health-check aggregation. Admin only.

pub fn unregister_health_contract( &mut self, contract: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> Result<(), MonitoringError>

Unregister a contract from health-check aggregation. Admin only.

pub fn try_unregister_health_contract( &mut self, contract: <<MonitoringContract as ContractEnv>::Env as Environment>::AccountId, ) -> MessageResult<Result<(), MonitoringError>>

Unregister a contract from health-check aggregation. Admin only.

pub fn get_health_contracts( &self, ) -> Vec<<<MonitoringContract as ContractEnv>::Env as Environment>::AccountId>

Get list of registered health-check contracts.

pub fn try_get_health_contracts( &self, ) -> MessageResult<Vec<<<MonitoringContract as ContractEnv>::Env as Environment>::AccountId>>

Get list of registered health-check contracts.